added temp, city, country, description to weather coponent and cleaned up app.js
